+++ internal/provider/job_definition_resource.go
@@ -6,8 +6,8 @@ import (
 	"strconv"
 	"terraform-provider-trocco/internal/client"
 	"terraform-provider-trocco/internal/client/parameter"
-	params "terraform-provider-trocco/internal/client/parameter/job_definitions"
-	filterParameters "terraform-provider-trocco/internal/client/parameter/job_definitions/filter"
+	params "terraform-provider-trocco/internal/client/parameter/job_definition"
+	filterParameters "terraform-provider-trocco/internal/client/parameter/job_definition/filter"
 	"terraform-provider-trocco/internal/provider/model"
 	job_definitions "terraform-provider-trocco/internal/provider/model/job_definition"
 	"terraform-provider-trocco/internal/provider/model/job_definition/filter"
