+++ internal/provider/model/job_definition/filter/filter_string_transform.go
@@ -2,7 +2,7 @@ package filter
 
 import (
 	filterEntities "terraform-provider-trocco/internal/client/entity/job_definition/filter"
-	"terraform-provider-trocco/internal/client/parameter/job_definitions/filter"
+	"terraform-provider-trocco/internal/client/parameter/job_definition/filter"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
