+++ internal/provider/model/job_definition/filter/filter_unix_time_conversion.go
@@ -2,7 +2,7 @@ package filter
 
 import (
 	"terraform-provider-trocco/internal/client/entity/job_definition/filter"
-	filter2 "terraform-provider-trocco/internal/client/parameter/job_definitions/filter"
+	filter2 "terraform-provider-trocco/internal/client/parameter/job_definition/filter"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
