+++ internal/provider/model/job_definition/input_option/mysql.go
@@ -2,7 +2,7 @@ package input_options
 
 import (
 	"terraform-provider-trocco/internal/client/entity/job_definition/input_option"
-	input_options2 "terraform-provider-trocco/internal/client/parameter/job_definitions/input_options"
+	input_options2 "terraform-provider-trocco/internal/client/parameter/job_definition/input_options"
 	"terraform-provider-trocco/internal/provider/model"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
