+++ internal/provider/model/job_definition/input_option/parser/xml_parser.go
@@ -2,7 +2,7 @@ package parser
 
 import (
 	job_definitions "terraform-provider-trocco/internal/client/entity/job_definition"
-	params "terraform-provider-trocco/internal/client/parameter/job_definitions"
+	params "terraform-provider-trocco/internal/client/parameter/job_definition"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
