+++ internal/provider/model/job_definition/job_definition_notification.go
@@ -2,7 +2,7 @@ package job_definitions
 
 import (
 	job_definitions "terraform-provider-trocco/internal/client/entity/job_definition"
-	params "terraform-provider-trocco/internal/client/parameter/job_definitions"
+	params "terraform-provider-trocco/internal/client/parameter/job_definition"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
